refactor(l1proto/envelope-fmt): organize into builder and parser
