+++ crates/l1proto/envelope-fmt/Cargo.toml
@@ -9,3 +9,6 @@ workspace = true
 [dependencies]
 bitcoin.workspace = true
 thiserror.workspace = true
+
+[features]
+test-utils = []
