+++ crates/l1proto/envelope-fmt/src/parser/mod.rs
@@ -1,0 +1,8 @@
+mod envelope;
+mod errors;
+
+pub use envelope::{
+    SignedEnvelopeLeaf, parse_envelope_container, parse_envelope_payload,
+    parse_multi_envelope_payloads, parse_signed_envelope_leaf,
+};
+pub use errors::EnvelopeParseError;
